+++ tests/fake_furnace/test_fake_furnace.py
@@ -19,6 +19,7 @@ def test_ramp():
     c = yaqc.Client(39426)
     c.set_ramp_time(0)
     c.set_position(0)
+    time.sleep(0.1)
     assert c.get_position() == 0
     c.set_ramp_time(1)
     c.set_position(100)