+++ yaqd-core/yaqd_core/_has_limits.py
@@ -15,24 +15,34 @@ class HasLimits(HasPosition, IsDaemon):
         self._out_of_limits = config["out_of_limits"]
 
     def get_limits(self) -> List[float]:
-        assert self._state["hw_limits"][0] < self._state["hw_limits"][1]
-        config_limits = self._config["limits"]
-        assert config_limits[0] < config_limits[1]
-        out = [
-            max(self._state["hw_limits"][0], config_limits[0]),
-            min(self._state["hw_limits"][1], config_limits[1]),
-        ]
+        # wrapper for client
+        return self.limits
+
+    @property
+    def limits(self) -> List[float]:
+        # for internal use
+        return self._joint_limit(self._state["hw_limits"], self._config["limits"])
+
+    @classmethod
+    def _joint_limit(cls, *limits: List[float]):
+        mins, maxes = [*zip(*limits)]
+        assert all([mini < maxi for mini, maxi in zip(mins, maxes)])
+        out = [max(*mins), min(*maxes)]
         assert out[0] < out[1]
         return out
 
     def in_limits(self, position: float) -> bool:
-        low, upp = self.get_limits()
+        return self._in_limits(position)
+
+    def _in_limits(self, position):
+        # for internal use
+        low, upp = self.limits
         return low <= position <= upp
 
     def set_position(self, position: float) -> None:
-        if not self.in_limits(position):
+        if not self._in_limits(position):
             if self._out_of_limits == "closest":
-                low, upp = self.get_limits()
+                low, upp = self.limits
                 if position > upp:
                     position = upp
                 elif position < low:
@@ -40,5 +50,5 @@ class HasLimits(HasPosition, IsDaemon):
             elif self._out_of_limits == "ignore":
                 return
             else:
-                raise ValueError(f"{position} not in ranges {self.get_limits()}")
+                raise ValueError(f"{position} not in ranges {self.limits}")
         super().set_position(position)